+++ src/main/java/WaguriException.java
@@ -1,0 +1,7 @@
+package waguri;
+
+public class WaguriException extends Exception {
+        public WaguriException(String message) {
+            super(message);
+        }
+    }
