+++ src/main/java/waguri/storage/Storage.java
@@ -1,8 +1,13 @@
+package waguri.storage;
+
+import waguri.task.Deadline;
+import waguri.task.Event;
+import waguri.task.Task;
+import waguri.task.Todo;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.Scanner;
