+++ src/main/java/waguri/task/Event.java
@@ -1,3 +1,5 @@
+package waguri.task;
+
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
