+++ src/main/java/waguri/task/Todo.java
@@ -1,3 +1,5 @@
+package waguri.task;
+
 public class Todo extends Task {
     public Todo(String description){
         super(description, false);
