+++ src/main/java/waguri/ui/Ui.java
@@ -1,3 +1,5 @@
+package waguri.ui;
+
 import java.util.Scanner;
 
 public class Ui {
